requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
amqp==5.3.1
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.11.0
async-timeout==5.0.1
attrs==25.4.0
billiard==4.2.4
celery==5.6.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
dj-database-url==3.0.1
Django==5.2.9
django-cors-headers==4.9.0
django-filter==25.2
django-jazzmin==3.0.1
django-lifecycle==1.2.4
django-ninja==1.5.1
django-stubs==5.2.8
django-stubs-ext==5.2.8
djangorestframework==3.16.1
drf-yasg==1.21.11
exceptiongroup==1.3.0
fastapi==0.121.0
frozenlist==1.8.0
gunicorn==23.0.0
h11==0.16.0
idna==3.11
inflection==0.5.1
kombu==5.6.1
multidict==6.7.0
packaging==25.0
pillow==12.0.0
prompt_toolkit==3.0.52
propcache==0.4.1
psycopg2==2.9.11
psycopg2-binary==2.9.11
pydantic==2.12.4
pydantic-settings==2.11.0
pydantic_core==2.41.5
pyserial==3.5
pyTelegramBotAPI==4.29.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
redis==7.0.1
requests==2.32.5
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.4
starlette==0.49.3
tomli==2.3.0
types-PyYAML==6.0.12.20250915
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.11.0
yarl==1.22.0
